Fix any new warnings from our own code related to upgrade to XCode 5.1 or 64-bit impact on NS(U)Integer

+++ Shelby.tv/Shelby.tv/SPVideoPlayer.m
@@ -639,8 +639,8 @@ NSString * const kShelbySPVideoAirplayDidEnd = @"spAirplayDidEnd";
     NSInteger to = (NSInteger)CMTimeGetSeconds(toTime);
     [ShelbyAPIClient postUserWatchedFrame:self.videoFrame.frameID
                                completely:complete
-                                     from:[NSString stringWithFormat:@"%01d", from]
-                                       to:[NSString stringWithFormat:@"%01d", to]];
+                                     from:[NSString stringWithFormat:@"%01ld", (long)from]
+                                       to:[NSString stringWithFormat:@"%01ld", (long)to]];
     _lastPlaybackUpdateIntervalEnd = toTime;
 }
